Decrease the aspect of the whole display
The whole aspect of the portal was huge, large buttons, texts and images. The whole aspect is now smaller, results are displayed in four columns instead of three (to avoid too much white on both sides), images are cropped to be sure they are all in the same size and videos titles, in list, can be on two lines max.