Make better use of material colors and elements
Menus (navigation and links) were implemented using md-toolbar element which uses the primary and accent colors as background colors. Primary and accent palettes are intention colors and must not be mistaken with background colors. It now uses the md-navbar element which comes with the new version of Angular Material and which is more appropriate to build a menu.
To continue with colors, themes was confusing (2 Material themes) and palettes of colors were mixing grey and blue for the default Portal theme. Only one Material theme survived: the Material default theme, no more references to the inverse Material theme. Default Portal theme now overwrites default theme palettes (primary, accent and background). For better understanding, all colors of the palettes are defined with documentation about related elements. As OpenVeo Portal has a blue dominant color, primary and accent palettes extend the Material blue palette.
Finally connection button was duplicated for mobile and larger devices to be inside the menu while on mobile. Connection / deconnection buttons are now in the same place for all devices.