Improve login / logout buttons to be more generic
A simple connection button, leading to cas service, was displayed for
CAS authentication. Authentication button now opens a dialog to choose
the authentication which prepares the interface for multiple
authentication mechanisms.
Also the logout button was displayed next to the name of the
authenticated user prefixed by "Hello". No more "Hello", the logout
button is now under a user menu.