Add a back office to OpenVeo Portal
Users connected to the portal had access to the
front office with or without restrictions
depending on their groups. The super administrator
has access to all front office videos. Super
administrator can now access /be URI to access the
back office. Only the super administrator is able
to access the back office.

This first version of the back office includes a
dashboard page with a block indicating the version
of the project and if a new version is available.
A navigation menu gives the possibility to change
interface language, return to the front office or
disconnect.

The back office is responsive and respects
Material specification.

The back office has no relation with the front
office, it uses AngularJS components structure and
can be compiled using "grunt build-admin". When
developping use "grunt watch:admin" to
automatically compile SASS and JavaScript files on
modifications.

JavaScript files are auto-detected and should not
be added to the conf.json file.