Dissociate ARIA label and help in opa-info-button
Back office opa-info-button component wasn't
making any distinction between the ARIA label and
the tooltip message. ARIA label was also used as
the tooltip message. As it is two different things
ARIA label and tooltip message are now separated.
opa-help sets the tooltip message while opa-label
sets the accessibility message.

Also opa-info-button had a default value for the
opa-label attribute which was modifying the value
of the opa-label attribute. This default
value was using opaTranslate filter and thus had a
dependency on it. It now has no default value and
opaTranslate is no longer a dependency of
opa-info-button, making it more independent.

Finally opa-label and opa-help were set as
one-way binding properties even if it can't be
anything else but Strings. Consequently both
opa-label and opa-tooltip are now simple Strings.