Clarify opa-toolbar accessibility messages
opa-toolbar was using confused action properties
to set accessibility messages. It now uses
uniformely the action property "accessibility" to
set the accessibility message of an action.

Also list action menu items used label as
the accessibility message. It is now possible to
set the accessibility message independently of the
label.