Add back office pages information button
Back office pages on small devices usually have
the title in the toolbar, not in the page as it is
on large devices. Consequently it wasn't possible
to add an information button next to the title. It
is now possible to add an information message for
each page of the back office. Actually this
button is only rendered on small devices. The
information button is displayed in the toolbar.