Prefix all web service routes
Web service routes and other routes (serving
html, CSS, JavaScript and so on) were mixed
together. Consequently collisions appeared between
AngularJS routes and web service routes. All
routes returning JSON content are now considered
as web service routes and are prefixed by /ws for
unsecured routes and /be/ws for routes requiring
authentication. As the web service URL is not
hardcoded in the client code and the AngularJS
back office has a /be base for its requests it
wasn't possible to make secure routes on /ws/be.