Revise Controller/Model/Provider/Database
Controller was using Models to save / fetch datas
from a Database. Model should not be used as
effective Objects but only as a general concept in
the same level as controllers and views.
Consequently the notion of models has been
completely removed from the project. Models code
has been reparted into controllers and providers.
Models removed a new system appears:
Controller / Provider / Database, but a Provider
could use different storages, not only databases.
Thus Database now extends Storage which leads to
the final system Controller / Provider / Storage.
The controller receives and validates the request,
perform necessary access controls and call
providers to manipulate resources. The provider
validates the resource format and ask the storage
to manipulate the resource in a precised location.

It is also now possible to update all user
properties not just email and password.