Refactor web service and cache mechanism
Requests to OpenVeo web service were using the
OpenVeo web service client accessible through the
context. Some requests were cached, other not.
FilterCache and VideoCache were used to perform
requests to OpenVeo web service. Like this
requests to OpenVeo web service were performed
either using the web service client when no cache
was needed or through the FilterCache and
VideoCache. This architecture has been changed to
be more easy to use and avoid confusions. Requests
to OpenVeo web service are now performed only
through the OpenVeoProvider. An instance of the
OpenVeoProvider is available in the context. The
OpenVeoProvider uses a Storage like any other
providers. OpenVeoProvider is capable of managing
cache for its requests with a default cache of 10
seconds for all requests with the possibility to
override it by request. OpenVeoProvider also
offers methods to delete documents from cache.