Add support for dateTime properties filters
It wasn't possible to add a custom property of
type "Date time" in the search engine. It is now
possible to filter videos using custom properties
of type "Date time" by entering a date. Search
will be made on all videos with a date between
this date at midnight and next day at midnight.