Fix back end compilation
Back end compilation was failing because back end
libraries were retrieved using the missing
"mountPath" property of the back end library
definition.

Back end libraries are not automatically loaded
by the server, thus back end libraries definition
must be a reference to the real libraries
definitions (top level "libraries" property of
the conf.json file).
The property "name" of a library stands for the
real name of the library and is used to find the
real path of the library. The property "mountPath"
holds the fake name of the library mounted by the
server. Consequently libraries have to be
referenced by the property "mountPath" instead of
the property "name".