server: add the ability to save search settings
Settings with id "search" can now be saved and
given back using SettingsController. For now only
one property is allowed in search settings:
"pois" which holds a boolean that can be used by
the client to search in points of interest by
default or not.
Also the search settings are available in
openVeoPortalSettings global variable in front
client.