client.front: change the display of videos results
Videos results in ALL VIDEOS page were displayed
through four columns with, for each video, its
title, thumbnail, date and number of views.

The display has been changed to be able to
integrate texts matching the query. When query is
made using the "Search by keyword" field, texts
containing the searched query are now displayed
for each video. Query can be found in title,
description but also in titles and descriptions of
chapters and tags. Part of texts containing the
query are now emphasised, truncated and displayed.

To display these new information and also the old
information, results are now displayed in two
columns. Each cell contains the thumbnail and the
details next to it.

"he" library is used to remove all HTML entities
from texts coming from WYSIWYG.

Emphasising part of texts matching the query will
not always work as the algorithm is not as
powerfull as the MongoDB algorithm which performs
the search. One of the caveat is accents, MongoDB
ignore accents while OpenVeo Portal doesn't.