client.front: improve truncate algorithm
Truncate algorithm wasn't able to take more words
before or after the text to be closer to the
limit, when removing words it was only removing
one word from both side of the emphasised text.
Also it was necessary to specify the number of
expected words around the emphasised text which
was hard to predict.
The algorithm is now capable to use more words
after or before the emphasised text and
automatically set the number of words around the
emphasised text.

Also the algorithm is now capable to emphasis text
inside a word.