feature: add Vimeo live
It is now possible to add a Vimeo live to the LIVE
page as it was for Youtube, Wowza and Vodalys.

Note that the LIVE page has been totally
factorized and is now part of the AngularJS
application, live is displayed directly in the
LIVE page and not in an iframe anymore.
Consequently there is no more particular
translation files for Wowza player, translations
have been added to the theme translations. Custom
themes MUST be updated!