client.admin: fix settings save
Saving settings was sometimes failing with generic
error notification and code 772 on server side
(wrong parameters).

It was due to the sub formulars
searchSettingsController and
liveSettingsController sometimes calling
settingsController with invalid settings because,
sometimes, AngularJs didn't have the time to
validate the sub formular when a field changed and
was calling settingsController while the formular
was invalid.